Autorização e Autenticação com Identity

+++ RazorPagesPizza/Pages/Pizza.cshtml.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using RazorPagesPizza.Models;
@@ -5,39 +6,44 @@ using RazorPagesPizza.Services;
 
 namespace RazorPagesPizza.Pages
 {
-    public class PizzaModel : PageModel
-    {
-        public List<Pizza> pizzas = new();
+	[Authorize]
+	public class PizzaModel : PageModel
+	{
+		public bool IsAdmin => HttpContext.User.HasClaim("IsAdmin", bool.TrueString);
+		public List<Pizza> pizzas = new();
 
-        [BindProperty]
-        public Pizza NewPizza { get; set; } = new();
+		[BindProperty]
+		public Pizza NewPizza { get; set; } = new();
 
-        public void OnGet()
-        {
-            pizzas = PizzaService.GetAll();
-        }
+		public void OnGet()
+		{
+			pizzas = PizzaService.GetAll();
+		}
 
-        public IActionResult OnPost()
-        {
-            if (!ModelState.IsValid)
-            {
-                return Page();
-            }
-            PizzaService.Add(NewPizza);
-            return RedirectToAction("Get");
-        }
+		public IActionResult OnPost()
+		{
+			if(!IsAdmin) return Forbid();
+			
+			if (!ModelState.IsValid)
+			{
+				return Page();
+			}
+			PizzaService.Add(NewPizza);
+			return RedirectToAction("Get");
+		}
 
-        public IActionResult OnPostDelete(int id)
-        {
-            PizzaService.Delete(id);
-            return RedirectToAction("Get");
-        }
+		public IActionResult OnPostDelete(int id)
+		{
+			if(!IsAdmin) return Forbid();
+			PizzaService.Delete(id);
+			return RedirectToAction("Get");
+		}
 
-        public string GlutenFreeText(Pizza pizza)
-        {
-            if (pizza.IsGlutenFree)
-                return "Gluten Free";
-            return "Not Gluten Free";
-        }
-    }
+		public string GlutenFreeText(Pizza pizza)
+		{
+			if (pizza.IsGlutenFree)
+				return "Gluten Free";
+			return "Not Gluten Free";
+		}
+	}
 }
